backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pymssql==2.3.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic==0.40.0
httpx==0.28.1
redis==5.2.1
apscheduler==3.10.4
